Phase 2 B2: extract shared helpers from truth_api.py into _truth_helpers.py
Move ~450 lines of constants, caches, and pure utility functions
(formatting, JSONL I/O, frontmatter parsing, FTS tokenization, path
helpers, briefing scoring) into a new _truth_helpers module. This
breaks the leaf-dependency chain so that future governance_api and
truth_queries facades can import helpers without circular deps.

truth_api.py: 6102 → 5725 lines (−377 net, after adding import block).

Made-with: Cursor
